Refactor: (BRD-107) EmailFormatPolicy -> EmailManager 재명명
