+++ app/src/main/kotlin/com/ttasjwi/board/system/Main.kt
@@ -1,6 +1,5 @@
 package com.ttasjwi.board.system
 
-import com.ttasjwi.board.system.common.annotation.component.DomainPolicy
 import com.ttasjwi.board.system.common.annotation.component.DomainService
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan
@@ -16,7 +15,6 @@ import org.springframework.scheduling.annotation.EnableAsync
         type = FilterType.ANNOTATION,
         classes = [
             DomainService::class,
-            DomainPolicy::class,
         ]
     )]
 )
