--- app/src/main/kotlin/com/ttasjwi/board/system/common/annotation/component/DomainPolicy.kt
@@ -1,8 +1,0 @@
-package com.ttasjwi.board.system.common.annotation.component
-
-/**
- * 도메인 정책, 규칙을 나타내는 마커 어노테이션입니다.
- */
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.CLASS)
-annotation class DomainPolicy
