+++ app/src/main/kotlin/com/ttasjwi/board/system/member/application/mapper/RegisterMemberCommandMapper.kt
@@ -6,7 +6,7 @@ import com.ttasjwi.board.system.common.logging.getLogger
 import com.ttasjwi.board.system.common.time.TimeManager
 import com.ttasjwi.board.system.member.application.dto.RegisterMemberCommand
 import com.ttasjwi.board.system.member.application.usecase.RegisterMemberRequest
-import com.ttasjwi.board.system.member.domain.policy.EmailFormatPolicy
+import com.ttasjwi.board.system.member.domain.service.EmailManager
 import com.ttasjwi.board.system.member.domain.service.NicknameManager
 import com.ttasjwi.board.system.member.domain.service.PasswordManager
 import com.ttasjwi.board.system.member.domain.service.UsernameManager
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Component
 
 @Component
 internal class RegisterMemberCommandMapper(
-    private val emailFormatPolicy: EmailFormatPolicy,
+    private val emailManager: EmailManager,
     private val passwordManager: PasswordManager,
     private val usernameManager: UsernameManager,
     private val nicknameManager: NicknameManager,
@@ -52,7 +52,7 @@ internal class RegisterMemberCommandMapper(
             exceptionCollector.addCustomExceptionOrThrow(NullArgumentException("email"))
             return null
         }
-        return emailFormatPolicy.validate(email)
+        return emailManager.validate(email)
             .getOrElse {
                 log.warn(it)
                 exceptionCollector.addCustomExceptionOrThrow(it)
