+++ app/src/main/kotlin/com/ttasjwi/board/system/member/application/service/EmailSendApplicationService.kt
@@ -1,18 +1,18 @@
 package com.ttasjwi.board.system.member.application.service
 
 import com.ttasjwi.board.system.member.application.usecase.EmailSendUseCase
-import com.ttasjwi.board.system.member.domain.policy.EmailFormatPolicy
+import com.ttasjwi.board.system.member.domain.service.EmailManager
 import com.ttasjwi.board.system.member.domain.service.EmailSender
 import org.springframework.stereotype.Service
 
 @Service
 internal class EmailSendApplicationService(
-    private val emailFormatPolicy: EmailFormatPolicy,
+    private val emailManager: EmailManager,
     private val emailSender: EmailSender,
 ) : EmailSendUseCase {
 
     override fun sendEmail(address: String, subject: String, content: String) {
-        val email = emailFormatPolicy.validate(address).getOrThrow()
+        val email = emailManager.validate(address).getOrThrow()
         emailSender.send(email, subject, content)
     }
 }
