--- app/src/main/kotlin/com/ttasjwi/board/system/member/domain/policy/EmailFormatPolicy.kt
@@ -1,5 +1,0 @@
-package com.ttasjwi.board.system.member.domain.policy
-
-interface EmailFormatPolicy {
-    fun validate(email: String): Result<String>
-}
