+++ app/src/main/kotlin/com/ttasjwi/board/system/member/domain/service/EmailManager.kt
@@ -1,0 +1,5 @@
+package com.ttasjwi.board.system.member.domain.service
+
+interface EmailManager {
+    fun validate(email: String): Result<String>
+}
