+++ app/src/main/kotlin/com/ttasjwi/board/system/member/domain/service/impl/EmailManagerImpl.kt
@@ -1,12 +1,12 @@
-package com.ttasjwi.board.system.member.domain.policy.impl
+package com.ttasjwi.board.system.member.domain.service.impl
 
-import com.ttasjwi.board.system.common.annotation.component.DomainPolicy
+import com.ttasjwi.board.system.common.annotation.component.DomainService
 import com.ttasjwi.board.system.member.domain.exception.InvalidEmailFormatException
-import com.ttasjwi.board.system.member.domain.policy.EmailFormatPolicy
+import com.ttasjwi.board.system.member.domain.service.EmailManager
 import org.apache.commons.validator.routines.EmailValidator
 
-@DomainPolicy
-class EmailFormatPolicyImpl : EmailFormatPolicy {
+@DomainService
+class EmailManagerImpl : EmailManager {
 
     private val apacheEmailValidator = EmailValidator.getInstance()
