+++ app/src/test/kotlin/com/ttasjwi/board/system/member/application/mapper/EmailVerificationCommandMapperTest.kt
@@ -6,7 +6,7 @@ import com.ttasjwi.board.system.common.exception.ValidationExceptionCollector
 import com.ttasjwi.board.system.common.time.fixture.TimeManagerFixture
 import com.ttasjwi.board.system.common.time.fixture.appDateTimeFixture
 import com.ttasjwi.board.system.member.application.usecase.EmailVerificationRequest
-import com.ttasjwi.board.system.member.domain.policy.fixture.EmailFormatPolicyFixture
+import com.ttasjwi.board.system.member.domain.service.fixture.EmailManagerFixture
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
@@ -23,7 +23,7 @@ class EmailVerificationCommandMapperTest {
     fun setup() {
         timeManagerFixture = TimeManagerFixture()
         emailVerificationCommandMapper = EmailVerificationCommandMapper(
-            emailFormatPolicy = EmailFormatPolicyFixture(),
+            emailManager = EmailManagerFixture(),
             timeManager = timeManagerFixture
         )
     }
@@ -66,7 +66,7 @@ class EmailVerificationCommandMapperTest {
     @Test
     @DisplayName("이메일 포맷이 유효하지 않을 때 예외 발생")
     fun testInvalidEmailFormat() {
-        val email = EmailFormatPolicyFixture.ERROR_EMAIL
+        val email = EmailManagerFixture.ERROR_EMAIL
         val code = "1234"
         val request = EmailVerificationRequest(email, code)
