+++ app/src/test/kotlin/com/ttasjwi/board/system/member/application/mapper/EmailVerificationStartCommandMapperTest.kt
@@ -6,8 +6,8 @@ import com.ttasjwi.board.system.common.locale.fixture.LocaleManagerFixture
 import com.ttasjwi.board.system.common.time.fixture.TimeManagerFixture
 import com.ttasjwi.board.system.common.time.fixture.appDateTimeFixture
 import com.ttasjwi.board.system.member.application.usecase.EmailVerificationStartRequest
-import com.ttasjwi.board.system.member.domain.policy.EmailFormatPolicy
-import com.ttasjwi.board.system.member.domain.policy.fixture.EmailFormatPolicyFixture
+import com.ttasjwi.board.system.member.domain.service.EmailManager
+import com.ttasjwi.board.system.member.domain.service.fixture.EmailManagerFixture
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.*
 import java.util.*
@@ -16,17 +16,17 @@ import java.util.*
 class EmailVerificationStartCommandMapperTest {
 
     private lateinit var commandMapper: EmailVerificationStartCommandMapper
-    private lateinit var emailFormatPolicy: EmailFormatPolicy
+    private lateinit var emailManager: EmailManager
     private lateinit var localeManagerFixture: LocaleManagerFixture
     private lateinit var timeManagerFixture: TimeManagerFixture
 
     @BeforeEach
     fun setup() {
-        emailFormatPolicy = EmailFormatPolicyFixture()
+        emailManager = EmailManagerFixture()
         localeManagerFixture = LocaleManagerFixture()
         timeManagerFixture = TimeManagerFixture()
         commandMapper = EmailVerificationStartCommandMapper(
-            emailFormatPolicy = emailFormatPolicy,
+            emailManager = emailManager,
             timeManager = timeManagerFixture,
             localeManager = localeManagerFixture,
         )
@@ -48,7 +48,7 @@ class EmailVerificationStartCommandMapperTest {
         @Test
         @DisplayName("이메일이 형식이 유효하지 않으면 예외가 발생한다")
         fun testInvalidFormatEmail() {
-            val request = EmailVerificationStartRequest(email = EmailFormatPolicyFixture.ERROR_EMAIL)
+            val request = EmailVerificationStartRequest(email = EmailManagerFixture.ERROR_EMAIL)
 
             val exception = assertThrows<CustomException> { commandMapper.mapToCommand(request) }
             assertThat(exception.debugMessage).isEqualTo("이메일 포맷 예외 - 픽스쳐")
