+++ app/src/test/kotlin/com/ttasjwi/board/system/member/application/processor/EmailAvailableProcessorTest.kt
@@ -5,7 +5,7 @@ import com.ttasjwi.board.system.common.message.fixture.MessageResolverFixture
 import com.ttasjwi.board.system.member.application.dto.EmailAvailableQuery
 import com.ttasjwi.board.system.member.domain.model.Member
 import com.ttasjwi.board.system.member.domain.model.fixture.memberFixture
-import com.ttasjwi.board.system.member.domain.policy.fixture.EmailFormatPolicyFixture
+import com.ttasjwi.board.system.member.domain.service.fixture.EmailManagerFixture
 import com.ttasjwi.board.system.member.domain.service.fixture.MemberStorageFixture
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
@@ -23,9 +23,9 @@ class EmailAvailableProcessorTest {
     @BeforeEach
     fun setup() {
         val memberStorageFixture = MemberStorageFixture()
-        val emailFormatPolicy = EmailFormatPolicyFixture()
+        val emailManagerFixture = EmailManagerFixture()
         processor = EmailAvailableProcessor(
-            emailFormatPolicy = emailFormatPolicy,
+            emailManager = emailManagerFixture,
             memberFinder = memberStorageFixture,
             messageResolver = MessageResolverFixture(),
             localeManager = LocaleManagerFixture()
@@ -44,7 +44,7 @@ class EmailAvailableProcessorTest {
         @Test
         @DisplayName("이메일 포맷이 유효하지 않을 때, 이메일 포맷이 유효하지 않다는 결과를 반환한다.")
         fun testInvalidFormat() {
-            val query = EmailAvailableQuery(email = EmailFormatPolicyFixture.ERROR_EMAIL)
+            val query = EmailAvailableQuery(email = EmailManagerFixture.ERROR_EMAIL)
 
             val result = processor.checkEmailAvailable(query)
