+++ app/src/test/kotlin/com/ttasjwi/board/system/member/application/service/EmailVerificationStartApplicationServiceTest.kt
@@ -6,7 +6,7 @@ import com.ttasjwi.board.system.common.time.fixture.appDateTimeFixture
 import com.ttasjwi.board.system.member.application.mapper.EmailVerificationStartCommandMapper
 import com.ttasjwi.board.system.member.application.processor.EmailVerificationStartProcessor
 import com.ttasjwi.board.system.member.application.usecase.EmailVerificationStartRequest
-import com.ttasjwi.board.system.member.domain.policy.fixture.EmailFormatPolicyFixture
+import com.ttasjwi.board.system.member.domain.service.fixture.EmailManagerFixture
 import com.ttasjwi.board.system.member.domain.service.fixture.EmailVerificationCreatorFixture
 import com.ttasjwi.board.system.member.domain.service.fixture.EmailVerificationEventCreatorFixture
 import com.ttasjwi.board.system.member.domain.service.fixture.EmailVerificationStartedEventPublisherFixture
@@ -27,7 +27,7 @@ class EmailVerificationStartApplicationServiceTest {
         timeManagerFixture = TimeManagerFixture()
         applicationService = EmailVerificationStartApplicationService(
             commandMapper = EmailVerificationStartCommandMapper(
-                emailFormatPolicy = EmailFormatPolicyFixture(),
+                emailManager = EmailManagerFixture(),
                 localeManager = LocaleManagerFixture(),
                 timeManager = timeManagerFixture
             ),
